requirements.txt
torch
torchvision
lightning
opencv-python
hydra-core
numpy
tensorboardx
torchmetrics
ax-platform
notify-py
grad-cam